Bot cinema
When the user writes start, the bot greets the user and offers a choice of 30 genres. After the user has chosen a genre, he will be sent the best movie of this genre according to the version of the Movie Poster website. If the user wrote something else, the bot will answer: "I didn't understand you"